Fix what clang-tidy 23 found, and the two package gaps behind it
With a working analyser the job produced three findings, two of them real.

An implicit copy per loop iteration: QJsonArray's iterator yields
QJsonValueConstRef, so binding it to `const QJsonValue&` materialises a
temporary every time. Four such loops now take `const auto&`. clang-tidy saw
only one because analysis stopped at the first error; the rest were the same
mistake.

A signed operand in a bitwise expression: a quint8 digest byte promotes to int
before the OR against a quint64 accumulator. It is widened explicitly now.

The third, portability-avoid-pragma-once, fires on all forty headers.
`#pragma once` is this project's convention and is supported by every compiler
it builds with, so the check is disabled with that reason recorded rather than
churning every header for a portability this project does not need.

The sanitizer job failed for a different reason than it appeared to:

    No tests were found!!!

Its test preset filters to unit, integration and gui, so --label-regex package
matched nothing and noTestsAction correctly failed the job -- the guard against
a suite silently discovering zero tests, working. Sanitizers cover the
application, not the shipped artefact, and the specification's matrix asks
only for sanitized tests there, so that row no longer runs the package gate.

The macOS bundle inventory answered the question the last commit asked. It
contained libQt6Sql and the Cocoa, image-format and style plugins, but no
sqldrivers directory at all, so the application could not open its own
database. The SQLite driver plugin is now installed there, the same way the
Wayland platform plugin is on Linux -- each only where the deployment helper
leaves it out, so the plugins it does deploy keep the load paths it rewrote.

+++ src/domain/cullfinch/domain/AssociationPolicy.cpp
@@ -314,7 +314,9 @@ AssociationResult StemAssociationResolver::resolve(const CollectionId& collectio
         const QByteArray digest = hash.result();
         quint64 revision = 0;
         for (int i = 0; i < 8 && i < digest.size(); ++i) {
-            revision = (revision << 8) | static_cast<quint8>(digest.at(i));
+            // Widen to the unsigned accumulator before the bitwise OR: a
+            // quint8 promotes to int, which would mix signedness.
+            revision = (revision << 8U) | static_cast<quint64>(static_cast<quint8>(digest.at(i)));
         }
         asset.membershipRevision = revision;
     }
